ui tests - presentation refactory
