+++ Presentation/Presenters/SignUp/SignUpPresenter.swift
@@ -27,17 +27,18 @@ public final class SignUpPresenter {
         if let message = validate(viewModel: viewModel) {
             alertView.showMessage(viewModel: AlertViewModel(title: "Falha na validação", message: message))
         } else {
-            let addAccountModel = AddAccountModel(name: viewModel.name!, email: viewModel.email!, password: viewModel.password!, passwordConfirmation: viewModel.passwordConfirmation!)
             
             loadingView.display(viewModel: LoadingViewModel(isLoading: true))
             
-            addAccount.add(addAccountModel: addAccountModel) { [weak self] result in
+            addAccount.add(addAccountModel: SignUpMapper.toAddAccountModel(viewModel: viewModel)) { [weak self] result in
                 guard let self = self else { return }
                 
                 switch result {
                 case .failure: self.alertView.showMessage(viewModel: AlertViewModel(title: "Erro", message: "Algo inesperado aconteceu. tente novamente em alguns instantes."))
-                case .success: break
+                case .success: self.alertView.showMessage(viewModel: AlertViewModel(title: "Sucesso", message: "Conta criada com sucesso."))
                 }
+                
+                self.loadingView.display(viewModel: LoadingViewModel(isLoading: false))
             }
         }
     }
@@ -71,17 +72,4 @@ public final class SignUpPresenter {
     }
 }
 
-public struct SignUpViewModel {
-    public var name: String?
-    public var email: String?
-    public var password: String?
-    public var passwordConfirmation: String?
-    
-    public init(name: String? = nil, email: String? = nil,
-                password: String? = nil, passwordConfirmation: String? = nil) {
-        self.name = name
-        self.email = email
-        self.password = password
-        self.passwordConfirmation = passwordConfirmation
-    }
-}
+
