+++ Presentation/Presenters/SignUp/SignUpViewModel.swift
@@ -1,0 +1,24 @@
+//
+//  SignUpViewModel.swift
+//  PresentationTests
+//
+//  Created by Israel Ermel on 02/09/20.
+//  Copyright © 2020 Israel Ermel. All rights reserved.
+//
+
+import Foundation
+
+public struct SignUpViewModel {
+    public var name: String?
+    public var email: String?
+    public var password: String?
+    public var passwordConfirmation: String?
+    
+    public init(name: String? = nil, email: String? = nil,
+                password: String? = nil, passwordConfirmation: String? = nil) {
+        self.name = name
+        self.email = email
+        self.password = password
+        self.passwordConfirmation = passwordConfirmation
+    }
+}
